Use typescript for configs
