+++ apps/web/eslint.config.js
@@ -1,6 +1,7 @@
 import tseslint from "typescript-eslint";
 import eslintPluginAstro from "eslint-plugin-astro";
 
+// TODO: fix this?
 import baseConfig from "lint/node.js";
 
 const config = tseslint.config(
